XmlTextReaderImpl: split ConvertAbsoluteUnixPathToAbsoluteUri in {.Windows,.Unix}
