+++ src/System.Private.Xml/src/System/Xml/Core/XmlTextReaderImpl.Windows.cs
@@ -1,0 +1,15 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System.Xml
+{
+    internal partial class XmlTextReaderImpl
+    {
+        private static string ConvertAbsoluteUnixPathToAbsoluteUri(string url, XmlResolver resolver)
+        {
+            return url;
+        }
+    }
+}
+
